fifth commit

+++ app/controllers/api/books_controller.rb
@@ -1,6 +1,6 @@
 module Api
   class BooksController < ApplicationController
-    # before_action :authenticate_user!, except: [:index]
+    before_action :authenticate_user!, except: [:index, :search]
     before_action :set_book, only: [:show, :update, :destroy]
 
     def index
@@ -16,28 +16,49 @@ module Api
       render json: { book: @book, author: @book.author }
     end
 
+    # for admin access
+    # def show
+    #   if (current_user && current_user.admin)
+    #     render json: { book: @book, author: @book.author }
+    #   else
+    #     render json: { message: "you have not access for this" }
+    #   end
+    # end
+
     def create
       @book = Book.new(book_params)
-      if @book.save
-        render json: { message: "Book added successfully" }, status: :ok
+      if (current_user && current_user.admin)
+        if @book.save
+          render json: { message: "Book added successfully" }, status: :ok
+        else
+          render json: { message: @book.errors.objects.first.full_message }, status: :bad_request
+        end
       else
-        render json: { message: @book.errors.objects.first.full_message }, status: :bad_request
+        render json: { message: "you are not access to it" }
       end
     end
 
     def update
-      if @book.update(book_params)
-        render json: { book: @book, author: @book.author }
+      if (current_user && current_user.admin)
+        if @book.update(book_params)
+          render json: { book: @book, author: @book.author }
+        else
+          render json: @book.errors
+        end
       else
-        render json: @book.errors
+        render json: { message: "you are not access to it" }
       end
     end
 
     def destroy
-      if @book.destroy
-        render json: { message: "Book Deleted successfully" }, status: :ok
+      if (current_user && current_user.admin)
+        if @book.destroy
+          render json: { message: "Book Deleted successfully" }, status: :ok
+        else
+          render json: { message: @book.errors.objects.first.full_message }, status: :bad_request
+        end
       else
-        render json: { message: @book.errors.objects.first.full_message }, status: :bad_request
+        render json: { message: "you are not access to it" }
       end
     end
